switch postgres to new cl import structure
