+++ splink/postgres/postgres_comparison_library.py
@@ -1,87 +1,8 @@
-from ..comparison_library import (
-    ArrayIntersectAtSizesComparisonBase,
-    DateDiffAtThresholdsComparisonBase,
-    DistanceFunctionAtThresholdsComparisonBase,
-    DistanceInKMAtThresholdsComparisonBase,
-    ExactMatchBase,
-    LevenshteinAtThresholdsComparisonBase,
+from .postgres_helpers.postgres_comparison_imports import (  # noqa: F401
+    array_intersect_at_sizes,
+    datediff_at_thresholds,
+    distance_function_at_thresholds,
+    distance_in_km_at_thresholds,
+    exact_match,
+    levenshtein_at_thresholds,
 )
-from .postgres_base import PostgresBase
-from .postgres_comparison_level_library import (
-    array_intersect_level,
-    datediff_level,
-    distance_function_level,
-    distance_in_km_level,
-    else_level,
-    exact_match_level,
-    levenshtein_level,
-    null_level,
-)
-
-
-class PostgresComparisonProperties(PostgresBase):
-    @property
-    def _exact_match_level(self):
-        return exact_match_level
-
-    @property
-    def _null_level(self):
-        return null_level
-
-    @property
-    def _else_level(self):
-        return else_level
-
-    @property
-    def _datediff_level(self):
-        return datediff_level
-
-    @property
-    def _array_intersect_level(self):
-        return array_intersect_level
-
-    @property
-    def _distance_in_km_level(self):
-        return distance_in_km_level
-
-    @property
-    def _levenshtein_level(self):
-        return levenshtein_level
-
-
-class exact_match(PostgresComparisonProperties, ExactMatchBase):
-    pass
-
-
-class distance_function_at_thresholds(
-    PostgresComparisonProperties, DistanceFunctionAtThresholdsComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return distance_function_level
-
-
-class levenshtein_at_thresholds(
-    PostgresComparisonProperties, LevenshteinAtThresholdsComparisonBase
-):
-    @property
-    def _distance_level(self):
-        return levenshtein_level
-
-
-class array_intersect_at_sizes(
-    PostgresComparisonProperties, ArrayIntersectAtSizesComparisonBase
-):
-    pass
-
-
-class datediff_at_thresholds(
-    PostgresComparisonProperties, DateDiffAtThresholdsComparisonBase
-):
-    pass
-
-
-class distance_in_km_at_thresholds(
-    PostgresComparisonProperties, DistanceInKMAtThresholdsComparisonBase
-):
-    pass
